+++ client/sass/_footer.scss
@@ -1,6 +1,5 @@
 .react-footer {
     background-color: #001A49;
-    margin-top: 2em;
     padding: 10px 0px;
     position: relative;
     bottom: 0px;
