+++ client/sass/_forms.scss
@@ -1,0 +1,17 @@
+.btn-lex {
+    background-color: #0057A8;
+    color: #fff;
+}
+
+.left-list {
+    border-right: 1px #BFBFBF solid;
+    li {
+        list-style: none;
+        padding: 1em 0em;
+        a {
+            font-size: 1.5em;
+            font-weight: normal;
+            color: #585858;
+        }
+    }
+}
