+++ client/sass/_layout.scss
@@ -2,6 +2,14 @@
     width: 95%;
 }
 
+.below-banner {
+    background-color: #f1f0f0;
+    color: #585858;
+    padding: 2em 0em 3em 0em;
+    font-family: 'Helvetica Neue', 'Helvetica', 'Arial', sans-serif;
+    font-weight: lighter;
+}
+
 .clearfix::after {
     content: "";
     clear: both;
