+++ client/sass/_navbar.scss
@@ -34,7 +34,6 @@ header {
     width: 100%;
     height: auto;
     display: inline-block;
-    padding-bottom: 1.5em;
 }
 
 .navbar-menu li {
